fix some python build bugs

+++ src/python/photon.py
@@ -23,6 +23,6 @@ def animate(i):
 
 
 ani = animation.FuncAnimation(
-    fig, animate, interval=1000/framerate, blit=True, save_count = framerate * totalTime)
+    fig, animate, interval=1000/framerate, blit=True, save_count = int(framerate * totalTime))
 ani.save(f"{outputFolder}photon.mp4")
 # plt.show()
